Revert "Pull Request #1: app_modes -> master"
