--- README.md
@@ -1,22 +1,0 @@
-## README ##
-
-## RasterPaint ##
-
-RasterPaint is a simple program approaching raster graphics. 
-
-## It allows user to: ##
-* Add points, lines and polygons (of different width and color),
-* List all the objects from the scene, edit and delete them,
-* Show grid (with given grid size) and different color,
-* Show or hide background on the scene.
-
-## Technology: ##
-* Logic: C#
-* Design: WPF (XAML)
-
-## Packages used: ##
-* [WriteableBitmapEx](https://writeablebitmapex.codeplex.com/)
-* [WpfExtendedToolkit](http://wpftoolkit.codeplex.com/)
-
-## Licence: ##
-[MIT Licence](https://opensource.org/licenses/MIT)
