+++ RasterPaint/RasterPaint.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 14.0.23107.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RasterPaint", "RasterPaint\RasterPaint.csproj", "{7AD482BF-E54A-4B0F-81DF-F7A20BEEFAB3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RasterPaintTests", "RasterPaintTests\RasterPaintTests.csproj", "{266AFEB3-13C9-4080-9AC0-EAE661E90B8C}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -17,10 +15,6 @@ Global
 		{7AD482BF-E54A-4B0F-81DF-F7A20BEEFAB3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7AD482BF-E54A-4B0F-81DF-F7A20BEEFAB3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7AD482BF-E54A-4B0F-81DF-F7A20BEEFAB3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{266AFEB3-13C9-4080-9AC0-EAE661E90B8C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{266AFEB3-13C9-4080-9AC0-EAE661E90B8C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{266AFEB3-13C9-4080-9AC0-EAE661E90B8C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{266AFEB3-13C9-4080-9AC0-EAE661E90B8C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
