+++ RasterPaint/RasterPaint/HelpWindow.xaml
@@ -1,0 +1,25 @@
+﻿<Window x:Class="RasterPaint.HelpWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:RasterPaint"
+        xmlns:system="clr-namespace:System;assembly=mscorlib"
+        mc:Ignorable="d"
+        Title="HelpWindow" Height="300" Width="300"
+        MinHeight="300" MinWidth="300"
+        ResizeMode="NoResize">
+    <Window.Resources>
+        <system:String x:Key="DeleteDescr">Usuwanie obiektu polega na sprawdzeniu, czy punkt należy do tzw. MyBoundary obiektu, czyli prostokąta ograniczonego największymi i najmniejszymi współrzędnymi. W przypadku kilku obiektów, które pasują do punktu, pojawi się kilka okien, z którego należy wybrać tylko te, które odpowiada obiektowi, który chcemy usunąć (będzie on podświetlony).</system:String>
+        <system:String x:Key="LeaveDescr">Wyjście poza obszar rysowania podczas rysowania sprawi, że wielokąt zostanie zamknięty. Nie dotyczy to punktów oraz linii (tylko co najmniej dwa wierzchołki składające się na obiekt.</system:String>
+    </Window.Resources>
+    <Grid>
+        <StackPanel>
+            <TextBlock Text="Help - pomoc do programu" Margin="5" TextAlignment="Center"></TextBlock>
+            <TextBlock Text="{StaticResource DeleteDescr}" TextWrapping="Wrap" TextAlignment="Justify" Margin="5" FontSize="11"/>
+            <TextBlock Text="{StaticResource LeaveDescr}" TextWrapping="Wrap" TextAlignment="Justify" Margin="5" FontSize="11"/>
+        </StackPanel>
+        
+        <Button Content="Close" Width="50" Click="ButtonBase_OnClick" HorizontalAlignment="Center" VerticalAlignment="Bottom" Margin="5"></Button>
+    </Grid>
+</Window>
