+++ RasterPaint/RasterPaint/ListWindow.xaml.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Shapes;
+
+namespace RasterPaint
+{
+    /// <summary>
+    /// Interaction logic for ListWindow.xaml
+    /// </summary>
+    public partial class ListWindow : Window
+    {
+        public ListWindow(List<MyObject> moList)
+        {
+            InitializeComponent();
+            Objects.ItemsSource = new ObservableCollection<MyObject>(moList);
+        }
+    }
+}
