+++ RasterPaint/RasterPaint/MyObject.cs
@@ -8,7 +8,7 @@ using System.Windows;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 
-namespace RasterPaint.Objects
+namespace RasterPaint
 {
     abstract public class MyObject
     {
@@ -25,8 +25,7 @@ namespace RasterPaint.Objects
         public abstract MyObject Clone();
         public abstract void UpdateBoundaries();
         public abstract void DrawObject(WriteableBitmap wb);
-        public abstract void EraseObject(List<MyObject> list, WriteableBitmap wb, Color c);
-        public abstract void HighlightObject(bool ifHighlight, WriteableBitmap wb, Color c);
-        public abstract bool IfPointCloseToBoundary(Point p);
+        public abstract void EraseObject(List<MyObject> list, WriteableBitmap wb);
+        public abstract void HighlightObject(bool ifHighlight, WriteableBitmap wb);
     }
 }
