--- RasterPaint/RasterPaint/Views/HelpWindow.xaml
@@ -1,28 +1,0 @@
-﻿<Window x:Class="RasterPaint.Views.HelpWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:RasterPaint"
-        xmlns:system="clr-namespace:System;assembly=mscorlib"
-        mc:Ignorable="d"
-        Title="HelpWindow" Height="450" Width="500"
-        MinHeight="300" MinWidth="300"
-        ResizeMode="NoResize">
-    <Window.Resources>
-        <BitmapImage x:Key="RasterPaintIcon" UriSource="/RasterPaint.ico"></BitmapImage>
-        <system:String x:Key="HelpString">Program RasterPaint zawiera kilkanaście funkcjonalności. Wybierz jeden z trzech trybów: Point, Line lub Polygon, by narysować odpowiednią figurę. Wybierz grubość jej krawędzi w zakresie od 0 do 5 i narysuj figurę. W przypadku rysowania w trybie „Polygon”, wyjście myszką poza obszar rysowania sprawi, że figura się „zamknie”. Tryb „Edit” pozwala zmienić kolor obiektu, jego wypełnienie (tylko Polygon), a także zmianę lokalizacji wierzchołków (tylko Polygon – złap za odpowiedni wierzchołek i przesuń go), oraz grubość obiektu – w locie. Move – pozwala na zmianę położenia obiektu, Remove – na jego usunięcie. Fill – wypełnia obszar (tylko Polygon), Clip – obcina obszarem prostokątnym. List – to lista wszystkich obiektów na scenie, Load / Save – pozwala na (de-)serializację całej sceny.</system:String>
-    </Window.Resources>
-    <Grid>
-        <StackPanel>
-            <TextBlock Text="Help - pomoc do programu" Margin="5" TextAlignment="Center"></TextBlock>
-            <TextBlock Text="{StaticResource HelpString}" TextWrapping="Wrap" TextAlignment="Justify" Margin="5" FontSize="14"/>
-            <Image Source="{StaticResource RasterPaintIcon}" Height="150" Width="150" Margin="5">
-                <Image.ToolTip>
-                    <ToolTip Content="Made by Grzegorz Czarnocki (c) 2015"></ToolTip>
-                </Image.ToolTip>
-            </Image>
-            <Button Content="Close" Width="50" Click="ButtonBase_OnClick" HorizontalAlignment="Center" VerticalAlignment="Bottom" Margin="5"></Button>
-        </StackPanel>
-    </Grid>
-</Window>
