--- RasterPaint/RasterPaint/Views/MainWindow.xaml
@@ -1,131 +1,0 @@
-﻿<Window x:Class="RasterPaint.Views.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:xctk="http://schemas.xceed.com/wpf/xaml/toolkit"
-        xmlns:local="clr-namespace:RasterPaint"
-        mc:Ignorable="d"
-        Title="RasterPaint" Height="350" Width="900"
-        MinHeight="350" MinWidth="900">
-
-    <Window.Resources>
-        <Style TargetType="Button" x:Key="ButtonStyle">
-            <Setter Property="Width" Value="55"/>
-            <Setter Property="Height" Value="25"/>
-            <Setter Property="Margin" Value="1"/>
-        </Style>
-
-        <Style TargetType="StackPanel" x:Key="InternalStackPanelStyle">
-            <Setter Property="Orientation" Value="Vertical"/>
-            <Setter Property="VerticalAlignment" Value="Center"/>
-            <Setter Property="Margin" Value="5"></Setter>
-        </Style>
-    </Window.Resources>
-
-    <Grid>
-
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="*"/>
-        </Grid.RowDefinitions>
-
-        <Grid Name="ImageGrid" Grid.Row="1" HorizontalAlignment="Stretch" VerticalAlignment="Stretch"
-                SizeChanged="ImageGrid_SizeChanged">
-            <Image Name="MyImage" VerticalAlignment="Top" HorizontalAlignment="Left"        
-                MouseLeftButtonDown="MyImage_ButtonDown"
-                MouseLeftButtonUp="MyImage_ButtonUp"
-                MouseLeave="MyImage_MouseLeave"
-                MouseMove="MyImage_OnMouseMove"/>
-        </Grid>
-
-        <StackPanel Grid.Row="0" Orientation="Horizontal" VerticalAlignment="Center" HorizontalAlignment="Center">
-            <StackPanel Style="{StaticResource InternalStackPanelStyle}">
-                <RadioButton GroupName="DrawingType" Content="Point" x:Name="PointRadioButton"
-                             Checked="DrawingType_Checked"/>
-                <RadioButton GroupName="DrawingType" Content="Line" x:Name="LineRadioButton"
-                            Checked="DrawingType_Checked"/>
-                <RadioButton GroupName="DrawingType" Content="Polygon" x:Name="PolygonRadioButton" IsChecked="True"
-                             Checked="DrawingType_Checked"/>
-                <xctk:IntegerUpDown Name="LineWidth" Height="25" Margin="5" Value="0" Minimum="0" Maximum="5" HorizontalContentAlignment="Center" VerticalContentAlignment="Center"
-                                ValueChanged="LineWidth_OnValueChanged"/>
-            </StackPanel>
-
-            <StackPanel Style="{StaticResource InternalStackPanelStyle}">
-                <Button Name="DrawGridButton" Content="Grid" Style="{StaticResource ButtonStyle}"
-                    Click="DrawGridButton_Click"/>
-
-                <xctk:IntegerUpDown Name="GridSize" Height="25" Margin="5" Value="25" Minimum="10" Maximum="50" HorizontalContentAlignment="Center" VerticalContentAlignment="Center"
-                                ValueChanged="GridSize_ValueChanged"/>
-            </StackPanel>
-
-            <StackPanel Style="{StaticResource InternalStackPanelStyle}">
-                <Label Content="Grid Color" HorizontalAlignment="Center" Height="25"/>
-
-                <xctk:ColorPicker Height="25" Width="75" Name="GridColorPicker" SelectedColor="Gray"
-                                  SelectedColorChanged="GridColor_OnSelectedColorChanged"/>
-
-                <Label Content="Bg. Color" HorizontalAlignment="Center"/>
-
-                <xctk:ColorPicker Height="25" Width="75" Name="BackgroundColorPicker" SelectedColor="LightYellow"
-                              SelectedColorChanged="BackgroundColor_OnSelectedColorChanged"/>
-            </StackPanel>
-
-            <StackPanel Style="{StaticResource InternalStackPanelStyle}">
-                <Label Content="Obj. Color" HorizontalAlignment="Center" VerticalAlignment="Bottom"/>
-
-                <xctk:ColorPicker Height="25" Width="75" Name="ObjectColorPicker" SelectedColor="DarkViolet"
-                                  SelectedColorChanged="ObjectColor_OnSelectedColorChanged"/>
-                
-                <Label Content="Fill Color" HorizontalAlignment="Center" VerticalAlignment="Bottom"/>
-
-                <xctk:ColorPicker Height="25" Width="75" Name="FillColorPicker" SelectedColor="CornflowerBlue"
-                                  SelectedColorChanged="FillColor_OnSelectedColorChanged"/>
-            </StackPanel>
-            
-            <StackPanel Style="{StaticResource InternalStackPanelStyle}">
-
-            </StackPanel>
-
-            <StackPanel Style="{StaticResource InternalStackPanelStyle}">
-                <Button Name="EditButton" Content="Edit" Style="{StaticResource ButtonStyle}"
-                    Click="EditButton_Click"/>
-
-                <Button Name="MoveButton" Content="Move" Style="{StaticResource ButtonStyle}"
-                    Click="MoveButton_Click"/>
-
-                <Button Name="RemoveButton" Content="Remove" Style="{StaticResource ButtonStyle}"
-                    Click="RemoveButton_Click"/>
-            </StackPanel>
-
-            <StackPanel Style="{StaticResource InternalStackPanelStyle}">
-                <Button Name="FillButton" Width="40" Content="Fill" Style="{StaticResource ButtonStyle}"
-                    Click="FillButton_Click"/>
-
-                <Button Name="ClipButton" Width="40" Content="Clip" Style="{StaticResource ButtonStyle}"
-                    Click="ClipButton_Click"/>
-            </StackPanel>
-
-            <StackPanel Style="{StaticResource InternalStackPanelStyle}">
-                <Button Name="HelpButton" Width="40" Content="Help" Style="{StaticResource ButtonStyle}"
-                    Click="HelpButton_Click"/>
-
-                <Button Name="ListButton" Width="40" Content="List" Style="{StaticResource ButtonStyle}"
-                    Click="ListButton_Click"/>
-            </StackPanel>
-
-            <StackPanel Style="{StaticResource InternalStackPanelStyle}">
-                <Button Name="LoadButton" Width="40" Content="Load" Style="{StaticResource ButtonStyle}"
-                        Click="LoadButton_OnClick"/>
-
-                <Button Name="SaveButton" Width="40" Content="Save" Style="{StaticResource ButtonStyle}"
-                        Click="SaveButton_OnClick"/>
-            </StackPanel>
-
-            <StackPanel Style="{StaticResource InternalStackPanelStyle}">
-                <Button Name="TestButton" Width="40" Content="Test" Style="{StaticResource ButtonStyle}"
-                        Click="TestButton_Click"/>
-            </StackPanel>
-        </StackPanel>
-    </Grid>
-</Window>
